+++ NEWS.md
@@ -10,11 +10,14 @@ We thank an anonymous reviewer for suggesting many of the improvements in this v
 * Changed example dataset `HEXACO` to also include cases with missing values.
 * `ipv_est` now also works for data in long format.
 * Example on how to add ggplot layers to an IPV chart to the vignette (this has been possible before, but should be emphasized)
-* Optional markers for the center distances in facet and nested charts to emphasize that the values are read at the border of the circles. 
+* Optional markers for the center distances in facet and nested charts emphasize that the values are read at the border of the circles (included by default).
+* When creating an `item_overview`, it is now possible to scale the font size (`size_font =`), specify the number of rows per facet to wrap overly wide charts (`wrap =`), scale the width (`width =`) and height (`height =`) of the file output, and select a subset of tests (`tests =`) and facets (`facets =`) for display.
 
 ## Bugfixes
 * resolved a warning message for nested charts that include tests without facets (no effect on output)
 * Function `rename` renamed to `relabel` to avoid conflicts with other packages.
+* Resolved multiple issues with inconsistent positioning of labels in `item_overview` charts and changed the default font to 'sans', since the consistent width of monospaced letters is no longer needed.
+* Changed several examples to include more appropriate models at the expense of runtime.
 
 # IPV 0.2.0